refactor: 파일명, 파일 구조 리팩토링 (#51)
* fix: 파일 네이밍 컨벤션에 맞춰 component 파일들 수정 -> PascalCase로 통일 #48

* refactor: websocket store 관련 파일 구조 변경, 파일명 kebab-case에 맞춰 수정 #46, #50
