+++ src/components/MenuButton.vue
@@ -26,7 +26,7 @@
 import ToggleButton from "primevue/togglebutton";
 import SplitButton from "primevue/splitbutton";
 import "primeicons/primeicons.css";
-import { useWebSocketStore } from "@/store/websocketStore";
+import { useWebSocketStore } from "@/store";
 import { ref } from "vue";
 
 const websocketStore = useWebSocketStore();
