+++ src/page/CapturePage.vue
