+++ src/store/index.ts
@@ -1,0 +1,7 @@
+import { createPinia } from "pinia";
+import { useWebSocketStore } from "./websocket";
+
+const pinia = createPinia();
+
+export { useWebSocketStore };
+export default pinia;
