+++ src/store/websocket/state.ts
@@ -1,0 +1,9 @@
+import { Message, StatMessage, PacketMessage } from "./types";
+
+export const state = () => ({
+  websocket: null as WebSocket | null,
+  isConnected: false,
+  message: {} as Message,
+  packetMessages: [] as PacketMessage[],
+  statMessage: {} as StatMessage,
+});
